REF: atmosphere model separated from environment
