+++ src/simulator/models/Models.jl
@@ -4,6 +4,7 @@ module Models
     include("state.jl")
     include("mass.jl")
     include("point_forces_moments.jl")
+    include("atmosphere.jl")
     include("environment.jl")
     include("aero_state.jl")
     include("controls.jl")
